Фикс+добавил поддержку LocalDate, LocalDateTime

+++ ormnext-core/src/main/java/ru/saidgadjiev/ormnext/core/query/criteria/impl/SelectStatement.java
@@ -183,7 +183,7 @@ public class SelectStatement<T> implements CriteriaStatement {
      * @param offset target offset
      * @return this for chain
      */
-    public SelectStatement<T> offset(int offset) {
+    public SelectStatement<T> offset(long offset) {
         this.offset = new Offset(offset);
 
         return this;
